Refactor loopback transport in valgrind tests.
The "ttrans" transport was shared by both tnpsrv and tnpfile
so pull it out to a separate file, ttrans.c and make it more
useful with an interface for performing rpc's on the transport
analagous to the client library.

misc/t12 (tnpfile under valgrind) became an expected pass after
these changes.