Fix test suite deadlock on Ubuntu 2.6.35 kernel
In the 'test/kern' tests, the mount system call was hanging because
'kconjoin', the parent of the client and server in this portion of
the test suite, was sloppy about closing duplicate file descriptors.
As a result, the kernel did not see an EOF on its connection to the
server when the server exited.  This addresses issue #27.

Also, the server would silently drop connections when deprecated 9P
ops were received.  The 2.6.35 kernel was an early stage of 9P2000.L
development where although 9P2000.L was negotiated, it was still sending
a TSTAT instead of a TGETATTR when handling the mount system call.
To make debugging this easier, deprecated protocol ops are regognized
(though not decoded) in the DEBUG_9P_PROTO output, and a message is
logged when one of these, or any other protocol error, is encountered.