Fix allsquash option.
Add test kern/t30 for allsquash and rework the kern and user runtest
script so mount and export directores are created in /tmp.

linbpfs was looking up users in auth and attach (and failing) even when
we were going to swap users to the squash user in the attach handler.
Add a new handler called srv->remapuser that can allocate the Npuser
for np_attach () and avoid the user lookup in the squash case.  Also,
in np_auth (), avoid doing a lookup if auth->startauth returns 0
indicating auth not required.  Rework some of this code for better
clarity and error handling.

Make 'squashuser' a config option (default = nobody).

Get rid of user.[ch] in and incorporate these functions directly into
the server, with improved error logging to catch more detail on the
"switch user" errors we are seeing in the logs.