Improve error handling for fsid switching
Add optional WT_FLAGS_SETFSID flag used at np_srv_create () time,
which turns setfsuid identity switching on and off at the libnpfs level.
This allows the current identity to be tracked per worker thread and
unnecessary identity changes avoided.

Fixes missing identity switches in diod's the lock/getlock handlers.

Failures of setfsuid, setfsgid, or setgroups are individually logged.

Verified that we are using the same error checking of setfsuid/setfsgid
as user space nfs server code, weak as it is.  The return values
documented in the man page do not seem to be accurate.