Rework libnpclient for test flexibility.
Abandon earlier misguided attempt to make libnpclient look more like
the C library.  Bring back direct access to the 9P protocol elements
for testing and re-add convenience functions implemented on top of
the core 9P functions.

Fix error handling so that it uses np_rerror () not errno for consistency
with the rest of the npfs code.

Update diodmount, diodexp various tests, and diod_auth to use the new API.