Replace diodexp with diodcat
The diodcat utility can cat any arbitrary file on a diod server
from user space, subject to the same authentication/authorization
rules as any diod client.

auto.diod is modified to use this utility to cat the diod server's
/proc/mounts to construct key matches, rather than cat the diodctl
exports pseudo-file using diodexp.  The upside of this change is
that a special purpose utility gets replaced with a general one,
and the only reason LLNL needs diodctl in diod 1.0 goes away.
The downside is that the server's /proc has to be made readable, and
requests for unexported file systems will trigger a (failing) mount
request instead of failing in the automouter key lookup.