Perform a late term abortion on diodctl.
As diodctl is no longer needed for release 1.0, it is removed.

The init script now starts 'diod' and the service is called diod not diodctl.

The config file 'diodctllisten' and 'diodlisten' variables are replaced
with 'listen'.

misc/t12, the sometimes failing valgrind test on npfile, is now gone.

Diod no longer has a -F option.

Diodmount no longer has a 'jobid' option, and it no longer has to
talk to diodctl of course.

The libnpfs srv_wait_timeout() and srv_shutdown() calls go away.

Upsides:
- Diodctl had no test coverage
- Npfile similarly had no coverage, and diodctl was the only user
- Kludgey method of testing whether a diod server had started
- Increased mount latency by requiring more protocol to be spoken
- Overall simplification of diod deployment.

Downsides:
- Not possible to start private servers on demand (for jobs or for users).
  Perhaps this can be addressed in the future through creative thread pool
  management within the single diod instance.
- Automounter integration is slightly less efficient in some error cases
  as described in a previous commit.