Don't lookup user in attach if auth already did it
Avoid back to back password file lookups between auth and attach.
Add np_afid2user () which checks if the attaching user is the same as
the auth user and then takes another reference on that Npuser and returns it.

Clean up log messages in auth and attach handlers in libnpfs.

Clean up log messages in diod_auth.c and make the code more resilient
to unexpected argument values.