Don't setfsgid to gid outside of user's sg set
Return EPERM to P9_LCREATE et al if the provided gid is not one of
the user's supplemental groups.

Without this check, a user could get elevated privilege in a server
running as root by attaching as themselves from user space and
issuing requests with arbitrary gids.  Add a testcase demonstrating
this.