Careful which thread destroys conn (issue 52)
When last reference is dropped, conn is destroyed which triggers
tpool cleanup.  If this reference is dropped by retirement of a request
instead of completion of the conn reader thread, the tpool cleanup
function will attempt to join with the worker thread that retired
the request, resulting in a pthread_join "deadlock avoided" error.
This race was causing misc/t10 to sporadically fail.

Have the conn reader thread explicitly handle the conn desruction
on termination.  Add a condition var so it can wait for retirement of
all outstanding reqs on the conn.