Bypass DAC check on root-authenticated connection (issue 53)
Calling setgroups () before every op turns out to be inadvisable since
the call can apply to all threads in some environments.  Instead, bypass
server-side discretionary access controls by setting CAP_DAC_OVERRIDE
capability on non-root fsid's when we know that v9fs is on the other
end and is already checking DACs.  This punts the supplemental group
check to the client.

We assume we have a "trusted" client if the connection was originally
authenticated as root.

None of this applies when the server is not running as root.