Move close to diod_clunk() (issue 71)
Close errors were masked as close occurred when the fid was
freed (via callback in the fidpool code), where there was no
opportunity to return an error.

Move this to diod_clunk () where it belongs and ensure errors
are returned properly, and that fid is removed regardless per
Plan9 clunk(5).