Add 'loosefid' flag to work around buggy client
v9fs does not currently handle a flushed Tclunk/Tremove properly.
It should behave as though these ops were never sent, meaning the fid
remains valid and cannot be reused.  However it unconditionally frees
them internally.  This must be fixed, however it is ueful to be able
to interoperate with older clients, so we create this workaround and
enable it by default.