Mmap first 'maxmmap' of a shared file
Instead of calling pread multiple times on a shared io context,
mmap up to 'maxmmap' of a file and call memcpy for read requests contained
in that area.  This is intended to help parallel executable load time.

The default value of 'maxmmap' is 4MB.  It can be changed via the
diod.conf maxmmap variable or the diod server -m,--maxmmap command
line option.