Acquire CAP_FOWNER (issue 98)
When v9fs opens a file for writing based on g+w permissions
(owner is different), it fails when it attempts to setattr the mtime
to NOW without changing the atime.  In cases where we have already
acquired the CAP_DAC_OVERRIDE, also acquire CAP_FOWNER to work around
this problem.