Fix possible race on dumpable flag (issue 105)
To ensure diod can always dump core, we added code to user.c to set
the dumpable flag after altering capabilities, fsuid, fsgid, or groups.
However, we only do this if the dumpable flag tested true before the
change.  This may be racy if another thread is executing the same code
and has cleared the dumpable flag before the test.