Avoid excessive network requests in npc_gets
Previous implementation of npc_gets was throwning away and re-reading
any characters after a \n.  Fix this so excess data is buffered in
the fid (as with readdir).

There was really no need for npc_closedir ().  Drop it and move buffer
cleanup to npc_clunk ().

This speeds up diodshowmount on test system with 1032 entries
in /proc/connections from 0.146s to 0.026s (real).