Require that "ctl" be explicitly exported.
This allows export options such as "ro" and "noauth" to be applied
to the ctl file system.

As part of the restructuring needed to implement this, the handling
of "ro" was moved to libnpfs so that it applies to both "ctl" and
the Npsrv abstract file system, in this case diod.

N.B. Tools that use ctl (like dtop) will no longer function until
"ctl" is added to the export list in diod.conf.