Diod: Lua detection fix and replacing deprecated function
lua_open was replaced with luaL_newstate, which succeeded it in Lua 5.2.
lua_open was already deprecated in Lua 5.1.
Also, in order to check for Lua during config, a test for luaL_newstate
is used, which works for both Lua 5.1 and 5.2, instead of testing for
lua_call, which does not work for Lua 5.2.

Signed-off-by: Leonid Bloch <leonid@daynix.com>