diod/ops.c: distinguish opened and unopened Fids in {set,get}attr
Previously the implementation always used Fid->path to get/set the
attributes. While this is fine for unopened fids, it creates problems
when the fid has been opened (and thus Fid->ioctx->fd exists):

 - it breaks the assumption that an opened file can still be
 interacted with, even if the path it has been opened from no longer
 exists (see #19).
 - Fid->ioctx->fd and Fid->path might refer to different files
 (e.g. Fid->path has been removed and recreated).

Consider:
	fd = open("file", O_RDONLY);
	remove("file);
	fstat(fd, &stat);

translated to 9P2000.L this is roughly:
	Twalk 0 1 "file"
	Tlopen 1 O_RDONLY
	Twalk 0 2 "file"
	Tremove 2
	Tgetattr 1 P9_GETATTR_ALL

since diod_getattr calls lstat("file") this failed with ENOENT.

Fix this by using the corresponding ioctx functions for opened fids,
which act on the fd instead of the path.

The corresponding testcases are in tests/user/t20

Fixes #19 #47