Fixes spec file for systemd build on CentOS 7.
In its current state, attempting to use the spec file to build a RPM package fails. First, the use of systemctl in the post-installation step needs a BuildRequires: systemd in the spec. Second, the double slash in the diod/Makefile.am made the verification phase of rpmbuild fail. This commit addresses both issues.