Remove empty README
The content of README was moved to README.md a while ago but an empty
README is confusingly still present.  The empty file is still there
since Makefile.am complains about it when autogen.sh is run if it's
missing.  Turn on the "foreign" option to reduce Automake's strictness
and remove the README.