Fix lua configure detection
We introduce a new configure switch called --with-lua-impl=,
which is used to store user defined and yet supported lua
implementation to use. If the parameter is not specified,
lua support is not built in.

For package maintainers this is a breaking fix, as it
changes the previous --with-lua-suffix, so watch out.

The previous implementation did not work on my system, no
matter what I tried, thus the detection and configuration
is now handled via PKG_HAVE_DEFINE_WITH_MODULES, which
seems to work fine for the various implementations, tested
with luajit, lua5.1 and lua5.4.