build: add configure --disable-config
Problem: config file support is silently disabled if lua
is not found.

Add configure --disable-config option.  By default, config file
support is enabled.  If enabled, it is a fatal error if lua
headers and libs are not found.  If disabled, lua is not required.

HAVE_CONFIG_FILE is defined in config.h if config file support
is enabled.

libdiod and tests now check for this instead of checking directly
for lua support.

Fixes #73