build: conditional use of gcc warning flags is broken
Problem: @GCCWARN@ is substituted even when the compiler is clang.

Import the AX_COMPILER_VENDOR macro and use it to test whether the
compiler vendor is clang or gnu.

Change the warning flags macro name to WARNING_CFLAGS, then
conditionally set it to different values for clang vs gnu.
A couple of the test programs unconditionally add -Wno-* options
that clang doesn't recognize, so tell clang not to complain about
unknown-warning-option errors.  For now, gcc is what's run in CI,
so let it fail on unknown-warning-option errors.

Tested with clang version 10.0.0-4ubuntu1.