diod: allow walk on open fids
Problem: ls -l on a mounted file system (tested with kernel 5.13.0)
fails with errors retrieving stat(2) information.

The kernel opens the directory fid and issues successive Treaddir
requests, and simultaneously walks from the directory fid to each
name in the directory, presumably to get a fid to call Tgetattr on.
This _seems_ reasonable, but on the other hand 9p walk(5) says
that walk from open fid is not permitted.

For now, revert "diod/ops.c: disallow walk on opened fids" since
this is a major problem.

This reverts commit 6749db36c8d297d22083def10c76ea1736a519b4.

Fixes #70