Signal that connection is done after all cleanup has finished
The connection thread is detached so we cannot join it to wait
for its compleation on shutdown. Use the conncount in server to
keep track of when all connection threads have actually finished.
To keep this safe we must wait to update this value until the
thread has done all of its cleanup.