libnpfs: fix off by one buffer check
Problem: buf_check_size() doesn't flag a buffer overflow if
the buffer is exactly full when the check is performed.

For example, a malformed Tlink with name size > 0 but no name causes
a segfault.

Fix off by one check so this and similar malformed requests
generate an error response.

Add unit test.

Fixes #109