testsuite: drop internal copy of dbench
Problem: dbench is vendored in the diod source tree, but it's a lot,
and is available as an ubuntu package.

Drop the dbench source code.
Modify the test script to skip the dbench test if unavailable.
Add the dbench package to github workflow in anticipation of running
the kernel tests later.