testsuite: set root mount propagation to private
Problem: the tests in tests/kern hang on systemd managed systems

Systemd sets the root file system propagation to shared, so the
private mounts created by the "kconjoin" test utilty are not
really private, and exiting the process that created them does
not cause an implicit unmount and diod does not exit.

Follow the lead of the unshare(1) command, which as of util-linux-2.27,
now sets / propgation recursively to private after unsharing it.

Fixes #61