testsuite: fix racy test looking for old bug
Problem: tests/kern/t41 fails in CI

This test fails if any dmesg output is produced while it runs - obviously
racy and inappropriate in cloud hosted CI testing.

Since the bug it is looking for is ancient, keep the test for its
possibly interesting file system workload, but drop the console check.