fix fsid in Rstatfs result
Problem: the remote fsid returned by Tstatfs is incorrect.

I botched the construction of fsid from fsid_t (an array of two variable
size integers) when building the response to Tstatfs.

Just use statvfs(3) instead, which returns the fsid as a single number.
It's more portable than statfs(2) so some ifdefs that cover BSD/linux
differences can go away.

Note that statfs(2) still has to be called to get the file system type
if the "statfs-passthrough" option is configured, but f_type seems to be
portable across platforms.

Fixes #32
Fixes #115