testsuite: convert capability test to TAP
Problem: the capability test run by tests/misc/t11 presents
results in an ad hoc manner.

Convert to TAP and co-locate with the code under test.

This test is skipped if it is not run as root.
Add scripts/check-root.sh to run unit tests as root that need it,
aborting if passwordless sudo does not work.