testsuite: convert config file test to TAP
Problem: the config file tests run by tests/misc/t06 and t08
present results in an ad hoc manner.

Convert to TAP and co-locate with the code under test.
Add this test to scripts/check-valgrind.sh as that might catch problems.

Drop the "tlua" program for checking whether diod was built with
config file support, as this is no longer needed.