libnpfs: replace kernel 9p.h header
Problem: during early diod development, a linux kernel header (`9p.h`) was
dropped into npfs to expedite the 9P2000.L port, but it is in a different
style and license than npfs.

To allow npfs to be redistributed under the original MIT license with the
9P2000.L changes, restore the orignal npfs data structures and definitions
and extend them for 9P2000.L.  As a consequence protocol operations and
flags are renamed, e.g.
  P9_TATTACH => Tattach
  P9_QTAUTH => Qtauth

Update users and tests.