testsuite: add locking unit test
Problem: there are no unit tests for Tlock and Tgetlock.

Add one that exercises the locking of whole files at the fid granularity.

In addition, demonstrate (simply) that byte range locking is not implemented.
We can get rid of the other tests that explore byte range locking, since the
lack of them is a design problem not a class of bugs that we need to keep testing,
as if they might work someday as if by magic.