diod: always run in the foreground
Problem: diod puts itself in the background by default, but this
can be confusing and unexpected when running the service manually.

In fact, given that systemd can manage daemons running in the
foreground, it is probably no longer useful at all.

Drop the --foreground option and never daemonize.
Update tests.