testsuite: fix test server TAP logging
Problem: some TAP unit tests emit debug output prefixed with "#: ",
which is legal TAP but looks funny.

Change diod_log_init() to not append ": " to the log prefix if it already
ends in a space, and change the test server to do that.

Also skip the log prefix if diod_log_init() is called with a NULL argument,
anticipating the next commit.