build: fix rdma configure logic
Problem: when --enable-rdmatrans is explicitly requested, failure to
configure it is non-fatal.

Make it fatal if the feature cannot be enabled.

Also rename "rdmatrans" to "rdma" and label it experimental in the
configure help output since it is not production ready.  For example,
if configured, it unconditionally tries to listen on a hard wired
address.