build: don't just warn if munge is missing
Problem: if munge is not found by configure, diod is silently
built without authentication support.

Make that a hard error unless configured with --disable-auth.
Also, use pkg-config to find munge.

Drop a GPL_LICENSED declaration in front of the munge.h include
directive.  The munge license has changed to LGPL and this is no
longer necessary.