testsuite: add diod-specific helper script
Problem: sharness tests will have to repeat a lot of boiler plate.

Add sharness.d/20-diod.sh which
- adds convenience variables pointing to common executables
- defines useful sharness prereqs
- provides functions to start/stop diod with standard test arguments

The "standard test arguments" including placing a diod.log in the
test directory, and setting $DIOD_SOCKET in the callers environment.
Note that placing the socket in the test directory doesn't always work
due to limitations on UNIX domain socket path lengths, so it is placed
in /tmp.