testsuite: create t1000-valgrind from old test
Problem: tests/misc/t15 uses the old test framework, but we are
converting to sharness.

Convert tests/misc/t15 => t1000-valgrind.t.

Remove tests/misc since this was the last test in that directory.

The valgrind suppressions no longer seem to be needed so for now,
start fresh without any.  Add them back if/when they are needed.