diod: keep --foreground option for compatibility
Commit 3dc9363f7bf8 ("diod: always run in the foreground") did what it says
on the tin and removed the functionality to daemonize. The old --foreground
option that disabled daemonization was also removed.

The breaks existing setups that were using foreground execution all along
via passing --foreground: The option no longer exists and instead the
full usage text is printed.

As not to break existing setups, let's keep the --foreground option
around for now and just turn it a no-op, given that it became the
default behavior.