testsuite: add diodrun --socketpair
Problem:  when mounting v9fs, it is more convenient to test with
diod on a unix domain socket than a passed file descriptor,
but test_under_diod assumes file descriptors.

Change diodrun so that it runs diod with a UNIX domain socket
by default, and uses passed file descriptors only when the
--socketpair option is specified.

Modify the server command line based on which mode is being used,
e.g. append:
  --listen=PATH --socktest
vs
  --rfdno=N --wfdno=N

Call setsid(2) on the server fork so that when diod is run under
sudo, the test output does not get mangled.

Update test_under_diod so there are now three methods:
  socketpair
    Add diodrun --socketpair option.
  unixsocket
    Add no options to diodrun.
  unixsocketroot
    Run server with sudo.