diod: drop incorrect assertion
Problem: Trename on a fid returned from Tcreate causes an assertion
failure.

assertion failure: diod_ioctx.c:431: path->ioctx == NULL

The assertion occurs when diod_rename() calls path_decref() on
the fid->path before replacing it with one containing the new name.
path_decref() insists that the fid not have an ioctx (e.g. not be open).

I think this is an incorrect assertion.
Remove it.

Add a test.

Fixes #93.