testsuite: fix POSIX ACL test
Problem: tests/kern/t43 fails with Operation not supported.

Set -oaccess=client,name=root for this test.
Remove t43 from the XFAIL list.

Fixes #106

Also, part of the test is commented out and contains a typo.
Fix the typo, update the expected output, and uncomment.
The test will be moving, but at least we will be starting
from a good place.