diod: fix alpine AF_UNIX getnameinfo failure
Problem: diod refuses to accept AF_UNIX connections on alpine linux.

The musl libc getnameinfo(3) fails with EAI_FAMILY when presented
with an AF_UNIX socket, unlike glibc which looks up the name as
"localhost" (NI_NUMERICHOST) or the local hostname.

Just set the client ID to one of the above when accepting an
AF_UNIX socket.