testsuite: fix allsquash mkdir test on alpine
Problem: a test that calls "mkdir -m 755 dir" as root on an
allsquash mount fails on alpine linux.

stracing "mkdir -m 755" shows that alpine calls
   mkdirat(AT_FDCWD, "/tmp/yyy", 0777)     = 0
   fchmodat(AT_FDCWD, "/tmp/yyy", 0755)    = 0

while glibc calls:
   mkdir("/tmp/zzz", 0755)                 = 0

Presumably the fchmodat is failing because, because root no longer
owns the file after the mkdirat due to the squashing.

In that test, perform the chmod separately as the squash user.