libdiod: simplify qid.type initialization
Problem: initializion of qid.type from a 'struct stat' or
'struct dirent' is confusing.

These functions set the qid.type "base type" to either
Qtfile, Qtdir, or Qtsymlink.  Just do that instead of
relying on the fact that Qtfile has a value of zero
and ORing in the other values.